+++ keyboards/idobao/montex/v1/readme.md
@@ -16,6 +16,6 @@ Then press the button labeled SW2 located between the 2 upper LEDs.
 
 Make example for this keyboard (after setting up your build environment):
 
-    make idobao/montex:default
+    make idobao/montex/v1:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
