+++ keyboards/idobao/montex/v1/v1.c
@@ -14,4 +14,4 @@
 * along with this program.  If not, see &lt;http://www.gnu.org/licenses/&gt;.
 */
 
-#include "montex.h"
+#include "v1.h"
